requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
Django==4.0.4
docx2txt==0.8
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
nltk==3.7
pygoogletranslation==2.0.6
PyPDF2==1.28.4
python-dotenv==0.20.0
regex==2022.4.24
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
sqlparse==0.4.2
tqdm==4.64.0
tzdata==2022.1
Unidecode==1.3.4
urllib3==1.26.9